% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dat.R
\docType{data}
\name{example_dat}
\alias{example_dat}
\title{A Simulated Example Dataset}
\format{
A list including:
\describe{
  \item{x}{Training input}
  \item{y}{Training output}
  \item{x_test}{Testing input}
  \item{y_test}{Tesing output}
}
}
\usage{
example_dat
}
\description{
A simulated examepl dataset includes training and testing
 data for twoPartDNN
}
\keyword{datasets}
