% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-classes.R
\docType{class}
\name{twoPartDNN-class}
\alias{twoPartDNN-class}
\title{An S4 class containing a twoPartDNN object}
\description{
An S4 class containing a twoPartDNN object
}
\section{Slots}{

\describe{
\item{\code{model.cls}}{A binary classification ensemble DNN model}

\item{\code{model.reg}}{A regression ensemble DNN model}
}}

\seealso{
\code{\link{dnnet-class}}\cr
\code{\link{ensemble_dnnet}}\cr
}
