% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-1-twoPartDNN.R
\name{twoPartDNN}
\alias{twoPartDNN}
\title{A Two-Part Ensemble DNN Model}
\usage{
twoPartDNN(object, ...)
}
\arguments{
\item{object}{A dnnetInput object}

\item{...}{Other parameters passing to ensemble_dnnet}
}
\value{
Returns a twoPartDNN object
}
\description{
A Two-Part Ensemble DNN Model
}
